requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
pandas==1.3.3
Pillow==8.3.2
requests==2.26.0
redis==4.0.2
rq==1.10.0
celery==5.1.2
psycopg2-binary==2.9.1
python-dotenv==0.19.1
gunicorn==20.1.0